Imagepicker added, NewsItems now a separate component. Newseditor improvements. Categorylogo now image. Refactor highlighteditormodal.
Added ImagePicker, that can be used to pick images from the uploaded files with ease.

Separated NewsItems to their own component, so that component can be used in the NewsEditor as well.

NewsEditor now shows the news as they would show up to the user, instead of an editor-specific view. Layout improved a lot in addition.

Changed logo to image in categories, making it more consistent with other objects like it.

Refactoring of HighlightEditorModal.